refactor(ui): migrate new user dropdown to all pages (#280)
