+++ app/dashboard/page.tsx
@@ -5,11 +5,6 @@ import { useForm } from "react-hook-form";
 import z from "zod";
 
 import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover"
-import {
   Card,
   CardDescription,
   CardHeader,
@@ -19,14 +14,11 @@ import { Switch } from "@/components/ui/switch";
 import { Button } from "@/components/ui/button"
 import { BetaBadge } from "@/components/ui/beta-badge";
 import { Input } from "@/components/ui/input";
-import { signOut } from "next-auth/react";
 import Link from "next/link";
 import { useState, useEffect } from "react";
 import {
   Plus,
   Trash2,
-  Settings,
-  LogOut,
   Grid3x3,
   Copy,
   Edit3,
@@ -54,7 +46,7 @@ import {
   DialogFooter,
 } from "@/components/ui/dialog";
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { ProfileDropdown } from "@/components/profile-dropdown";
 
 // Dashboard-specific extended types
 export type DashboardBoard = Board & {
@@ -290,9 +282,7 @@ export default function Dashboard() {
     }
   }
 
-  const handleSignOut = async () => {
-    await signOut();
-  };
+  
 
   if (loading) {
     return <FullPageLoader message="Loading dashboard..." />;
@@ -322,41 +312,9 @@ export default function Dashboard() {
               <Plus className="w-4 h-4" />
               <span className="hidden sm:inline">Add Board</span>
             </Button>
-            <Popover>
-               <PopoverTrigger asChild>
-                    <Avatar className="w-9 h-9 cursor-pointer">
-                    <div className="w-9 h-9 flex items-center justify-center rounded-full hover:bg-zinc-200 dark:hover:bg-zinc-700 transition-colors ">
-                    <AvatarImage className="w-7 h-7 rounded-full" src={user?.image || ""} alt={user?.name || ""} />
-                    <AvatarFallback className="w-8 h-8 flex items-center justify-center rounded-full text-zinc-900 dark:text-zinc-100 bg-blue-500 ">
-                        <span className="text-sm font-medium text-white">
-                          {user?.name
-                            ? user.name.charAt(0).toUpperCase()
-                            : user?.email?.charAt(0).toUpperCase()}
-                        </span>
-                      </AvatarFallback>
-                    </div>
-                    </Avatar>
-               </PopoverTrigger>
-               <PopoverContent className="w-80 bg-white dark:bg-zinc-900"> 
-                    <div>
-                        <p className="font-medium text-zinc-900 dark:text-zinc-100">
-                            {user?.name || user?.email}
-                        </p>
-                        <p className="text-xs text-zinc-500 dark:text-zinc-400">
-                            {user?.email}
-                        </p>
-                        <Link href={"/settings"} className="flex items-center mt-4 hover:bg-zinc-200 dark:hover:bg-zinc-800  pl-2 dark:hover:text-zinc-50 text-zinc-800 dark:text-zinc-400 rounded-md text-sm gap-2 py-2">
-                          <Settings size={19} />
-                            Settings
-                        </Link>
+              
 
-                        <div onClick={handleSignOut} className="flex items-center mt-2 group hover:bg-zinc-200 dark:hover:bg-zinc-800 pl-2 text-red-700 dark:hover:text-red-500 rounded-md cursor-pointer text-sm gap-2 py-2">
-                          <LogOut size={19} />
-                              Sign Out
-                        </div>
-                    </div>
-               </PopoverContent>
-            </Popover>
+            <ProfileDropdown user={user} />
           </div>
         </div>
       </nav>
