+++ app/settings/layout.tsx
@@ -2,13 +2,12 @@
 
 import { useState, useEffect, useCallback } from "react"
 import { useRouter, usePathname } from "next/navigation"
-import { User as UserIcon, Building2, ArrowLeft, Settings, LogOut, ChevronDown } from "lucide-react"
-import { Button } from "@/components/ui/button"
+import { User as UserIcon, Building2, ArrowLeft } from "lucide-react"
 import Link from "next/link"
 import { BetaBadge } from "@/components/ui/beta-badge"
-import { signOut } from "next-auth/react"
 import { FullPageLoader } from "@/components/ui/loader"
 import type { User } from "@/components/note"
+import { ProfileDropdown } from "@/components/profile-dropdown"
 
 export default function SettingsLayout({
   children,
@@ -17,7 +16,6 @@ export default function SettingsLayout({
 }) {
   const [user, setUser] = useState<User | null>(null)
   const [loading, setLoading] = useState(true)
-  const [showUserDropdown, setShowUserDropdown] = useState(false)
   const router = useRouter()
   const pathname = usePathname()
 
@@ -44,24 +42,6 @@ export default function SettingsLayout({
     fetchUserData()
   }, [fetchUserData])
 
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (showUserDropdown) {
-        const target = event.target as Element
-        if (!target.closest('.user-dropdown')) {
-          setShowUserDropdown(false)
-        }
-      }
-    }
-
-    document.addEventListener('mousedown', handleClickOutside)
-    return () => document.removeEventListener('mousedown', handleClickOutside)
-  }, [showUserDropdown])
-
-  const handleSignOut = async () => {
-    await signOut()
-  }
-
   if (loading) {
     return <FullPageLoader message="Loading settings..." />
   }
@@ -78,47 +58,7 @@ export default function SettingsLayout({
               <h1 className="text-xl sm:text-2xl font-bold text-blue-600 dark:text-blue-400 flex items-center gap-2">Gumboard <BetaBadge /></h1>
             </Link>
           </div>
-          <div className="relative user-dropdown">
-            <Button
-              onClick={() => setShowUserDropdown(!showUserDropdown)}
-              className="flex items-center space-x-2 text-foreground dark:text-zinc-100 hover:text-foreground dark:hover:text-white focus:outline-none focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 focus:ring-offset-2 rounded-md px-3 py-2"
-            >
-              <div className="w-8 h-8 bg-blue-500 dark:bg-blue-600 rounded-full flex items-center justify-center">
-                <span className="text-sm font-medium text-white">
-                  {user?.name ? user.name.charAt(0).toUpperCase() : user?.email?.charAt(0).toUpperCase()}
-                </span>
-              </div>
-              <span className="text-sm font-medium hidden md:inline">
-                {user?.name?.split(' ')[0] || 'User'}
-              </span>
-              <ChevronDown className="w-4 h-4 ml-1" />
-            </Button>
-
-            {showUserDropdown && (
-              <div className="absolute right-0 mt-2 w-48 bg-white dark:bg-zinc-900 rounded-md shadow-lg border border-gray-200 dark:border-zinc-800 z-50">
-                <div className="py-1">
-                  <div className="px-4 py-2 text-sm text-muted-foreground dark:text-zinc-400 border-b dark:border-zinc-800">
-                    {user?.email}
-                  </div>
-                  <Link
-                    href="/settings"
-                    className="flex items-center px-4 py-2 text-sm text-foreground dark:text-zinc-100 hover:bg-accent dark:hover:bg-zinc-800"
-                    onClick={() => setShowUserDropdown(false)}
-                  >
-                    <Settings className="w-4 h-4 mr-2" />
-                    Settings
-                  </Link>
-                  <Button
-                    onClick={handleSignOut}
-                    className="flex items-center w-full px-4 py-2 text-sm text-foreground dark:text-zinc-100 hover:bg-accent dark:hover:bg-zinc-800"
-                  >
-                    <LogOut className="w-4 h-4 mr-2" />
-                    Sign Out
-                  </Button>
-                </div>
-              </div>
-            )}
-          </div>
+          <ProfileDropdown user={user} />
         </div>
       </nav>
